♻️ refactor: adapt `users` to new app router
