+++ components/pages/users/users-create-page.component.tsx
@@ -1,61 +1,53 @@
+"use client"
+
 //Utility
 import axios from "axios"
 import Swal from "sweetalert2"
 import withReactContent from "sweetalert2-react-content"
-import Router, { useRouter } from "next/router"
+import { useRouter } from "next/navigation"
 import { FiLoader } from "react-icons/fi"
 import { useSession } from "next-auth/react"
+import checkPermGroup from "../../../lib/ui/check-perm-group"
+import checkFieldEmpty from "../../../lib/ui/check-field-empty"
 
 //React
-import { useRef, useState } from "react"
+import { useEffect, useRef, useState } from "react"
 
 //Components
-import Menu from "../../../../components/dashboard/menu"
+import Menu from "../../../components/dashboard/menu"
 
 //Interfaces
-import { PermissionGroup, User } from "../../../../interfaces"
-
-//Styles
-import "tippy.js/dist/tippy.css"
-import checkPermGroup from "../../../../lib/ui/check-perm-group"
-import checkFieldEmpty from "../../../../lib/ui/check-field-empty"
-import handleValueChange from "../../../../lib/ui/handle-value-change"
-import useSaveData from "../../../../hooks/use-save-data"
+import { PermissionGroup } from "../../../interfaces"
+import handleValueChange from "../../../lib/ui/handle-value-change"
+import useSaveData from "../../../hooks/use-save-data"
 
 //===============================================
 
-export async function getServerSideProps(req) {
-  const { slug } = req.query
-  const resPermGroupActionURI = `${process.env.BASE_URL}/api/v1/permission-groups?apikey=${process.env.API_KEY}`
-  const resPermissionGroups = await axios.get(resPermGroupActionURI)
-  const permissionGroups: PermissionGroup = await resPermissionGroups.data
-
-  const resUserActionURI = `${process.env.BASE_URL}/api/v1/users/_id/${slug}?apikey=${
-    process.env.API_KEY
-  }&secretkey=${process.env.SECRET_KEY}`
-  const resUser = await axios.get(resUserActionURI)
-  const user: User = await resUser.data
+export async function getServerSideProps() {
+  const res = await axios.get(`${process.env.BASE_URL}/api/v1/permission-groups?apikey=${process.env.API_KEY}`)
+  const permissionGroups: PermissionGroup = await res.data
 
   return {
     props: {
-      fetchedPermissionGroups: permissionGroups,
-      fetchedUser: user
+      fetchedPermissionGroups: permissionGroups
     }
   }
 }
 
-export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
-  const [formValues, setFormValues] = useState(fetchedUser)
-  const [currentPw] = useState(fetchedUser[0].password)
+export default function UsersCreatePage({ fetchedPermissionGroups }: { fetchedPermissionGroups: PermissionGroup[] }) {
+  const [formValues, setFormValues] = useState({})
+
   const [formErrors, setFormErrors] = useState({})
+
   const [showErrors, setShowErrors] = useState(false)
   const [showError, setShowError] = useState({})
-  const [isUpdateBtnClicked, setIsUpdateBtnClicked] = useState(false)
-  const saveData = useSaveData("USER", "UPDATE")
 
-  //routing
+  const [isCreateBtnClicked, setIsCreateBtnClicked] = useState(false)
+
+  const saveData = useSaveData("USER", "CREATE")
+
+  // router
   const router = useRouter()
-  const { slug } = router.query
 
   //sweetalert
   const resultSwal = withReactContent(Swal)
@@ -66,49 +58,91 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
   //Auth Session
   const { data: session } = useSession()
 
+  useEffect(() => {
+    //Add initial form values at first page render
+    if (Object.keys(formValues).length === 0) {
+      setFormValues({
+        username: "",
+        password: "",
+        email: "",
+        permission_group: ""
+      })
+    }
+
+    //Add initial form errors at first page render
+    if (Object.keys(formErrors).length === 0) {
+      setFormErrors({
+        username: "empty-field",
+        password: "empty-field",
+        email: "empty-field",
+        permission_group: "empty-field"
+      })
+    }
+  }, [])
+
   const submitData = async () => {
     if (Object.keys(formErrors).length > 0) {
       const formValuesWithErrors = Object.keys(formErrors)
       setShowErrors(true)
-      setIsUpdateBtnClicked(false)
+      setIsCreateBtnClicked(false)
       formRef.current[formValuesWithErrors[0]].focus()
     } else {
-      const result = await saveData({ formValues, slug, currentPw })
-      if (result.status === "success") {
+      const result = await saveData({ formValues })
+
+      const checkEmailExist: boolean = result.data.isEmailExist
+      const checkUsernameExist: boolean = result.data.isUsernameExist
+      if (!checkEmailExist && !checkUsernameExist) {
+        if (result.success) {
+          resultSwal
+            .fire({
+              title: `${result.data.message} Do you want to create another one?`,
+              icon: "success",
+              showDenyButton: true,
+              showCancelButton: false,
+              confirmButtonText: "Yes",
+              denyButtonText: "No",
+              confirmButtonColor: "#10B981"
+            })
+            .then((result) => {
+              if (result.isConfirmed) {
+                //Clear form and set default if user want create another one
+                ;(document.getElementById("user_create_form") as HTMLFormElement).reset()
+                setFormValues({ username: "", password: "", email: "", permission_group: "" })
+                setIsCreateBtnClicked(false)
+              } else if (result.isDenied) {
+                router.push("/dashboard/users")
+              }
+            })
+        } else {
+          resultSwal
+            .fire({
+              title: `${result.data.message} Do you want to change values and try again?`,
+              icon: "error",
+              showDenyButton: true,
+              showCancelButton: false,
+              confirmButtonText: "Yes",
+              denyButtonText: "No",
+              confirmButtonColor: "#10B981"
+            })
+            .then((result) => {
+              if (result.isDenied) {
+                router.push("/dashboard/users")
+              }
+            })
+        }
+      } else {
         resultSwal
           .fire({
-            title: `${result.message}. Do you want to continue editing it?`,
-            icon: "success",
-            showDenyButton: true,
-            showCancelButton: false,
-            confirmButtonText: "Yes",
-            denyButtonText: "No",
-            confirmButtonColor: "#10B981"
-          })
-          .then((result) => {
-            if (result.isDenied) {
-              void Router.push("/dashboard/users")
-            } else {
-              setIsUpdateBtnClicked(false)
-            }
-          })
-      } else if (result.status === "failed") {
-        resultSwal
-          .fire({
-            title: `${result.message}. Do you want to continue editing it?`,
+            title: `Already exist:\n ${checkEmailExist ? "Email" : ""}\n ${checkUsernameExist ? "Username" : ""}`,
             icon: "error",
-            showDenyButton: true,
+            showDenyButton: false,
             showCancelButton: false,
-            confirmButtonText: "Yes",
+            confirmButtonText: "OK",
             denyButtonText: "No",
             confirmButtonColor: "#10B981"
           })
-          .then((result) => {
-            if (result.isDenied) {
-              void Router.push("/dashboard/users")
-            } else {
-              setIsUpdateBtnClicked(false)
-            }
+          .then(() => {
+            setIsCreateBtnClicked(false)
           })
       }
     }
@@ -120,13 +154,13 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
         <Menu />
         <div className="grid grid-col-6 ml-10 mt-10 content-start place-content-around w-screen">
           <div className="col-start-1 col-end-2 ">
-            <h1 className="text-6xl font-josefin">Edit User </h1>
+            <h1 className="text-6xl font-josefin">New User</h1>
           </div>
           <div className="col-start-4 col-end-6 "></div>
           <div className="col-start-1 col-end-6 w-10/12 mt-10 ">
             {checkPermGroup(session, "admin") ? (
               <form action="#" id="user_create_form">
-                <div className="flex flex-col justify-center border-2 border-slate-200 p-10 ">
+                <div className="flex flex-col block justify-center border-2 border-slate-200 p-10 ">
                   <div className="w-11/12">
                     <label htmlFor="username" className="form-label inline-block mb-2 text-gray-700 text-xl">
                       Username
@@ -137,7 +171,7 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                       id="username"
                       name="username"
                       ref={(el) => (formRef.current[`username`] = el)}
-                      defaultValue={formValues[0].username}
+                      defaultValue={formValues["username"]}
                       onChange={(e) => {
                         handleValueChange(
                           formValues,
@@ -148,7 +182,7 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                           setShowError,
                           setFormValues,
                           e,
-                          "USER_EDIT"
+                          "USER"
                         )
                       }}
                       onBlur={(e) => {
@@ -172,7 +206,7 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                       id="email"
                       name="email"
                       ref={(el) => (formRef.current[`email`] = el)}
-                      defaultValue={formValues[0].email}
+                      defaultValue={formValues["email"]}
                       onChange={(e) => {
                         handleValueChange(
                           formValues,
@@ -183,7 +217,7 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                           setShowError,
                           setFormValues,
                           e,
-                          "USER_EDIT"
+                          "USER"
                         )
                       }}
                       onBlur={(e) => {
@@ -207,7 +241,8 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                       id="password"
                       name="password"
                       ref={(el) => (formRef.current[`password`] = el)}
-                      onChange={(e) => {
+                      defaultValue={formValues["password"]}
+                      onChange={(e) =>
                         handleValueChange(
                           formValues,
                           formErrors,
@@ -217,10 +252,19 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                           setShowError,
                           setFormValues,
                           e,
-                          "USER_EDIT"
+                          "USER"
                         )
+                      }
+                      onBlur={(e) => {
+                        checkFieldEmpty(formErrors, showError, setFormErrors, setShowError, e)
                       }}
+                      required
                     />
+                    {(showErrors || showError[`password`]) &&
+                      formErrors[`password`] &&
+                      formErrors[`password`] === "empty-field" && (
+                        <span className="text-red-500 font-bold">This field is required</span>
+                      )}
                   </div>
                   <div className="w-11/12 mt-5">
                     <label htmlFor="permission_group" className="form-label inline-block mb-2 text-gray-700 text-xl">
@@ -230,8 +274,8 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                       id="permission_group"
                       name="permission_group"
                       ref={(el) => (formRef.current[`permission_group`] = el)}
-                      defaultValue={formValues[0].permission_group}
-                      onChange={(e) => {
+                      defaultValue={formValues["permission_group"]}
+                      onChange={(e) =>
                         handleValueChange(
                           formValues,
                           formErrors,
@@ -241,13 +285,14 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                           setShowError,
                           setFormValues,
                           e,
-                          "USER_EDIT"
+                          "USER"
                         )
-                      }}
+                      }
                       onBlur={(e) => {
                         checkFieldEmpty(formErrors, showError, setFormErrors, setShowError, e)
                       }}
                       className="form-select form-select-lg mb-3 block w-full  px-4 py-2 text-xl font-normal text-gray-700 bg-white bg-clip-padding bg-no-repeat border border-solid border-gray-300  rounded transition ease-in-out m-0 focus:text-gray-700 focus:bg-white focus:border-slate-600 focus:outline-none">
+                      <option value="">Please choose a group</option>
                       {fetchedPermissionGroups.map((permissionGroup) => {
                         return <option value={`${permissionGroup.slug}`}>{`${permissionGroup.name}`}</option>
                       })}
@@ -264,21 +309,21 @@ export default function EditUser({ fetchedPermissionGroups, fetchedUser }) {
                   <div className=" w-11/12">
                     <button
                       type="button"
-                      data-testid="update_user_btn"
+                      data-testid="create_user_btn"
                       onClick={() => {
-                        setIsUpdateBtnClicked(true)
-                        submitData()
+                        setIsCreateBtnClicked(true)
+                        void submitData()
                       }}
                       className="mb-2 w-full inline-block px-6 py-2.5 bg-slate-700 text-white font-medium text-xs leading-normal uppercase rounded shadow-md hover:bg-slate-800 hover:shadow-lg focus:bg-slate-700 focus:shadow-lg focus:outline-none focus:ring-0 active:bg-slate-800 active:shadow-lg transition duration-150 ease-in-out">
-                      {isUpdateBtnClicked ? (
+                      {isCreateBtnClicked ? (
                         <span className="flex flex-row justify-center">
                           <FiLoader className="animate-spin text-2xl" />
-                          <span data-testid="update_user_btn_processing" className="mt-1 ml-3">
+                          <span data-testid="create_user_btn_processing" className="mt-1 ml-3">
                             Processing
                           </span>
                         </span>
                       ) : (
-                        "Update"
+                        "Create"
                       )}
                     </button>
                   </div>
